chore: avoid use of size_t in bbapi to be safe (#20916)
Use uint32_t in bbapi (as was already done in most places) to avoid any
platform dependent shenanigans.

+++ barretenberg/cpp/src/barretenberg/api/api_chonk.cpp
@@ -58,7 +58,7 @@ void ChonkAPI::prove(const Flags& flags,
     request.vk_policy = bbapi::parse_vk_policy(flags.vk_policy);
     std::vector<PrivateExecutionStepRaw> raw_steps = PrivateExecutionStepRaw::load_and_decompress(input_path);
 
-    bbapi::ChonkStart{ .num_circuits = raw_steps.size() }.execute(request);
+    bbapi::ChonkStart{ .num_circuits = static_cast<uint32_t>(raw_steps.size()) }.execute(request);
     info("Chonk: starting with ", raw_steps.size(), " circuits");
     for (const auto& step : raw_steps) {
         bbapi::ChonkLoad{
